+++ CHANGES.rst
@@ -18,6 +18,11 @@ Development
   a config scan.
 * ``config.register_authentication_policy`` accepts a list of contexts.
 
+0.2.1
+-----
+
+* Adjust requirements files and dependencies.
+
 0.2
 ---